Feat getMostLanguage from github & Fix user auth
